misc: move expect, expecthttp, handlerutils into utils
