+++ Dockerfile
@@ -2,6 +2,7 @@ FROM golang:1.14-alpine
 
 WORKDIR /go/src/app
 
+# TODO ADD psql
 RUN apk add --no-cache --update \
       openssl\
       git \
