+++ auth/apikey/apikey_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/DATA-DOG/go-txdb"
 	"github.com/cga1123/bissy-api/auth"
 	"github.com/cga1123/bissy-api/auth/apikey"
-	"github.com/cga1123/bissy-api/expect"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
 	"github.com/google/uuid"
 	"github.com/honeycombio/beeline-go/wrappers/hnysqlx"
 	"github.com/jmoiron/sqlx"
