+++ auth/github/client.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 
 	"github.com/cga1123/bissy-api/auth"
-	"github.com/cga1123/bissy-api/handlerutils"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/handlerutils"
 )
 
 // ClientState represents the state passed in by a client auth request,
