+++ auth/github/handlers.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/cga1123/bissy-api/auth"
 	"github.com/cga1123/bissy-api/auth/jwtprovider"
-	"github.com/cga1123/bissy-api/handlerutils"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/handlerutils"
 	"github.com/go-redis/redis/v8"
 	"github.com/gorilla/mux"
 	"github.com/honeycombio/beeline-go/wrappers/hnysqlx"
