+++ auth/github/handlers_test.go
@@ -12,10 +12,10 @@ import (
 	"github.com/cga1123/bissy-api/auth"
 	"github.com/cga1123/bissy-api/auth/github"
 	"github.com/cga1123/bissy-api/auth/jwtprovider"
-	"github.com/cga1123/bissy-api/expect"
-	"github.com/cga1123/bissy-api/expecthttp"
-	"github.com/cga1123/bissy-api/handlerutils"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
+	"github.com/cga1123/bissy-api/utils/expecthttp"
+	"github.com/cga1123/bissy-api/utils/handlerutils"
 	"github.com/google/uuid"
 	"github.com/gorilla/mux"
 	_ "github.com/lib/pq"
