+++ fsw.test.yml
@@ -5,8 +5,18 @@ triggers:
   - '**/*.go'
   cmd: |
     clear
+    echo ""
+    echo "* Drop"
+    psql -U bissy -d postgres -c "DROP DATABASE IF EXISTS bissy_test;"
+    echo " * Create"
+    psql -U bissy -d postgres -c "CREATE DATABASE bissy_test;"
+    echo " * Migrate"
+    migrate -path migrations -database $DATABASE_URL -verbose up
+    echo " * Test"
     go test -timeout=10s -coverprofile=cover.out -parallel=4 ./...
+    echo " * Coverage"
     go tool cover -html=cover.out -o coverage.html
+    echo " * Lint"
     golint ./...
     echo " * DONE"
   shell: true
