+++ querycache/datasource_handlers_test.go
@@ -5,11 +5,11 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/cga1123/bissy-api/expect"
-	"github.com/cga1123/bissy-api/expecthttp"
-	"github.com/cga1123/bissy-api/handlerutils"
 	"github.com/cga1123/bissy-api/querycache"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
+	"github.com/cga1123/bissy-api/utils/expecthttp"
+	"github.com/cga1123/bissy-api/utils/handlerutils"
 )
 
 func TestDatasourceCreate(t *testing.T) {
