+++ querycache/datasource_store_test.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/DATA-DOG/go-txdb"
-	"github.com/cga1123/bissy-api/expect"
 	"github.com/cga1123/bissy-api/querycache"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
 	"github.com/google/uuid"
 	"github.com/honeycombio/beeline-go/wrappers/hnysqlx"
 	"github.com/jmoiron/sqlx"
