+++ querycache/executor_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cga1123/bissy-api/expect"
 	"github.com/cga1123/bissy-api/querycache"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
 	"github.com/google/uuid"
 	_ "github.com/lib/pq"
 )
