+++ querycache/query_store_test.go
@@ -6,9 +6,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cga1123/bissy-api/expect"
 	"github.com/cga1123/bissy-api/querycache"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
 	"github.com/google/uuid"
 	"github.com/honeycombio/beeline-go/wrappers/hnysqlx"
 )
