+++ querycache/router_test.go
@@ -4,10 +4,10 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/cga1123/bissy-api/expect"
-	"github.com/cga1123/bissy-api/expecthttp"
-	"github.com/cga1123/bissy-api/handlerutils"
 	"github.com/cga1123/bissy-api/utils"
+	"github.com/cga1123/bissy-api/utils/expect"
+	"github.com/cga1123/bissy-api/utils/expecthttp"
+	"github.com/cga1123/bissy-api/utils/handlerutils"
 )
 
 func TestHome(t *testing.T) {
