+++ scripts/test.sh
@@ -8,7 +8,7 @@ dockerize \
   -wait tcp://postgres_test:5432 \
   -timeout 10s \
   -wait-retry-interval 1s \
-  migrate -path migrations -database $DATABASE_URL -verbose up
+  echo "External services are up!"
 
 echo " * run tests, watch"
 fswatch --config fsw.test.yml
